Organize into core components folder
