+++ dpc-portal/app/components/core/alert/component.html.erb
